BIG UPGRADE
This simulation can get the hit on the plastic scintillator from any generated particle, except for Optical Photons. It stores data about the position of detections and the particle's name. You can watch an histogram of it with the Ntuple and ROOT. Also, It has a random seed and works with macro file.